Add interop-data tests
