--- lang/csharp/src/apache/codec/Avro.Codec.sln
@@ -1,101 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-
-#  Licensed to the Apache Software Foundation (ASF) under one or more
-#  contributor license agreements.  See the NOTICE file distributed with
-#  this work for additional information regarding copyright ownership.
-#  The ASF licenses this file to You under the Apache License, Version 2.0
-#  (the "License"); you may not use this file except in compliance with
-#  the License.  You may obtain a copy of the License at
-#
-#      https://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32126.317
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.BZip2", "Avro.Codec.BZip2\Avro.Codec.BZip2.csproj", "{40201607-AA97-44FB-8353-6CDFA803FBBB}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.BZip2.Test", "Avro.Codec.BZip2.Test\Avro.Codec.BZip2.Test.csproj", "{ABEDA4C6-8024-48C1-AFC0-37190ABC659A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.Snappy", "Avro.Codec.Snappy\Avro.Codec.Snappy.csproj", "{C3F05185-2D65-4A63-B16D-6BE20A254C1F}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.Snappy.Test", "Avro.Codec.Snappy.Test\Avro.Codec.Snappy.Test.csproj", "{564A59B5-008D-4654-B63F-8EFAC8C45D10}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.XZ", "Avro.Codec.XZ\Avro.Codec.XZ.csproj", "{06396DF0-D656-4133-8BDC-EBF194D13AE4}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.XZ.Test", "Avro.Codec.XZ.Test\Avro.Codec.XZ.Test.csproj", "{4E8A6CCA-90DF-4350-B928-E6ADA5686883}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.Zstandard", "Avro.Codec.Zstandard\Avro.Codec.Zstandard.csproj", "{21E6C150-200F-484C-949F-C818BBE4F276}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.Zstandard.Test", "Avro.Codec.Zstandard.Test\Avro.Codec.Zstandard.Test.csproj", "{B93B1C26-C254-4C04-BB6B-E4FC8CCA9122}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.main", "..\main\Avro.main.csproj", "{E5B71984-FAEC-43E2-B2AC-4413D737C5CD}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Avro.Codec.Brotli", "Avro.Codec.Brotli\Avro.Codec.Brotli.csproj", "{9174B8A2-FBFF-4992-BF15-CE5729F695A7}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Avro.Codec.Brotli.Test", "Avro.Codec.Brotli.Test\Avro.Codec.Brotli.Test.csproj", "{9CBB2B45-DEE4-432D-8841-F8524F1F4F9A}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{40201607-AA97-44FB-8353-6CDFA803FBBB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{40201607-AA97-44FB-8353-6CDFA803FBBB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{40201607-AA97-44FB-8353-6CDFA803FBBB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{40201607-AA97-44FB-8353-6CDFA803FBBB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{ABEDA4C6-8024-48C1-AFC0-37190ABC659A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{ABEDA4C6-8024-48C1-AFC0-37190ABC659A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{ABEDA4C6-8024-48C1-AFC0-37190ABC659A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{ABEDA4C6-8024-48C1-AFC0-37190ABC659A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C3F05185-2D65-4A63-B16D-6BE20A254C1F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C3F05185-2D65-4A63-B16D-6BE20A254C1F}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C3F05185-2D65-4A63-B16D-6BE20A254C1F}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C3F05185-2D65-4A63-B16D-6BE20A254C1F}.Release|Any CPU.Build.0 = Release|Any CPU
-		{564A59B5-008D-4654-B63F-8EFAC8C45D10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{564A59B5-008D-4654-B63F-8EFAC8C45D10}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{564A59B5-008D-4654-B63F-8EFAC8C45D10}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{564A59B5-008D-4654-B63F-8EFAC8C45D10}.Release|Any CPU.Build.0 = Release|Any CPU
-		{06396DF0-D656-4133-8BDC-EBF194D13AE4}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{06396DF0-D656-4133-8BDC-EBF194D13AE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{06396DF0-D656-4133-8BDC-EBF194D13AE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{06396DF0-D656-4133-8BDC-EBF194D13AE4}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4E8A6CCA-90DF-4350-B928-E6ADA5686883}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4E8A6CCA-90DF-4350-B928-E6ADA5686883}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4E8A6CCA-90DF-4350-B928-E6ADA5686883}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4E8A6CCA-90DF-4350-B928-E6ADA5686883}.Release|Any CPU.Build.0 = Release|Any CPU
-		{21E6C150-200F-484C-949F-C818BBE4F276}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{21E6C150-200F-484C-949F-C818BBE4F276}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{21E6C150-200F-484C-949F-C818BBE4F276}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{21E6C150-200F-484C-949F-C818BBE4F276}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B93B1C26-C254-4C04-BB6B-E4FC8CCA9122}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B93B1C26-C254-4C04-BB6B-E4FC8CCA9122}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B93B1C26-C254-4C04-BB6B-E4FC8CCA9122}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B93B1C26-C254-4C04-BB6B-E4FC8CCA9122}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E5B71984-FAEC-43E2-B2AC-4413D737C5CD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5B71984-FAEC-43E2-B2AC-4413D737C5CD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5B71984-FAEC-43E2-B2AC-4413D737C5CD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5B71984-FAEC-43E2-B2AC-4413D737C5CD}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9174B8A2-FBFF-4992-BF15-CE5729F695A7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9174B8A2-FBFF-4992-BF15-CE5729F695A7}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9174B8A2-FBFF-4992-BF15-CE5729F695A7}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9174B8A2-FBFF-4992-BF15-CE5729F695A7}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9CBB2B45-DEE4-432D-8841-F8524F1F4F9A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9CBB2B45-DEE4-432D-8841-F8524F1F4F9A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9CBB2B45-DEE4-432D-8841-F8524F1F4F9A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9CBB2B45-DEE4-432D-8841-F8524F1F4F9A}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {32C95735-F692-433B-BCD2-A702CA701398}
-	EndGlobalSection
-EndGlobal
