+++ lang/csharp/src/apache/test/Avro.test.csproj
@@ -46,6 +46,13 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\codec\Avro.File.Snappy\Avro.File.Snappy.csproj" />
+    <ProjectReference Include="..\codec\Avro.File.BZip2\Avro.File.BZip2.csproj" />
+    <ProjectReference Include="..\codec\Avro.File.XZ\Avro.File.XZ.csproj" />
+    <ProjectReference Include="..\codec\Avro.File.Zstandard\Avro.File.Zstandard.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
   </ItemGroup>
